+++ packages/db/src/schema/index.ts
@@ -72,6 +72,8 @@ export { xTweetAnalytics } from "./x_tweet_analytics.js";
 export { mediaDrops } from "./media_drops.js";
 export { autoReplyConfig, autoReplyLog, autoReplySettings } from "./auto_reply.js";
 export { systemCrons } from "./system_crons.js";
+export { socialAccounts } from "./social_accounts.js";
+export { socialAutomations } from "./social_automations.js";
 export { partnerCompanies, partnerClicks, partnerSiteContent } from "./partners.js";
 export { affiliates } from "./affiliates.js";
 export { referralAttribution } from "./referral_attribution.js";